+++ Simulation/Aivika/Lattice/Generator.hs
@@ -3,7 +3,7 @@
 
 -- |
 -- Module     : Simulation.Aivika.Lattice.Generator
--- Copyright  : Copyright (c) 2016, David Sorokin <david.sorokin@gmail.com>
+-- Copyright  : Copyright (c) 2016-2017, David Sorokin <david.sorokin@gmail.com>
 -- License    : BSD3
 -- Maintainer : David Sorokin <david.sorokin@gmail.com>
 -- Stability  : experimental
@@ -30,8 +30,10 @@ instance MonadGenerator LIO where
   data Generator LIO =
     GeneratorLIO { generator01 :: LIO Double,
                    -- ^ the generator of uniform numbers from 0 to 1
-                   generatorNormal01 :: LIO Double
+                   generatorNormal01 :: LIO Double,
                    -- ^ the generator of normal numbers with mean 0 and variance 1
+                   generatorSequenceNo :: LIO Int
+                   -- ^ the generator of sequence numbers
                  }
 
   generateUniform = generateUniform01 . generator01
@@ -60,6 +62,8 @@ instance MonadGenerator LIO where
 
   generateDiscrete = generateDiscrete01 . generator01
 
+  generateSequenceNo = generatorSequenceNo
+
   newGenerator tp =
     case tp of
       SimpleGenerator ->
@@ -81,8 +85,14 @@ instance MonadGenerator LIO where
 
   newRandomGenerator01 g01 =
     do gNormal01 <- newNormalGenerator01 g01
+       gSeqNoRef <- liftIO $ newIORef 0
+       let gSeqNo =
+             do x <- liftIO $ readIORef gSeqNoRef
+                liftIO $ modifyIORef' gSeqNoRef (+1)
+                return x
        return GeneratorLIO { generator01 = g01,
-                             generatorNormal01 = gNormal01 }
+                             generatorNormal01 = gNormal01,
+                             generatorSequenceNo = gSeqNo }
 
 -- | Create a normal random number generator with mean 0 and variance 1
 -- by the specified generator of uniform random numbers from 0 to 1.