+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+Version 0.2
+-----
+
+* Implemented lazy references.
+
 Version 0.1.1
 -----